+++ cpf/input_types/XYFunctions.py
@@ -882,7 +882,7 @@ class XYDetector:
         all_colours = cm.seismic(np.arange(n_entries))
 
         if np.abs(maximum_value) > np.abs(minimum_value):
-            n_cut = np.int(
+            n_cut = np.int_(
                 (
                     (2 * maximum_value - (maximum_value - np.abs(minimum_value)))
                     / (2 * maximum_value)
@@ -892,7 +892,7 @@ class XYDetector:
             keep = n_entries - n_cut
             all_colours = all_colours[keep:]
         else:
-            n_cut = np.int(
+            n_cut = np.int_(
                 (
                     (
                         2 * np.abs(minimum_value)
